Update lc_matching.c
The formula for Q calculation is wrong. As stated in page 9 of https://www.mikrocontroller.net/attachment/473317/Crystal_Motional_Parameters.pdf (which this code is based from), **the fraction's denominator should be Reff**, which is 2* Zo+Rm.  However, the code uses Rm instead of Reff, which leads to absurdingly high Q values.

I noticed that when I was measuring a quartz crystal and VNA told me that the Q factor was over 77000 when it should be around 6800.